+++ client/404/index.html
@@ -1,0 +1,9 @@
+<html>
+<head>
+	Test
+</head>
+
+<body>
+	TEST TEST
+</body>
+</html>
