+++ points.json
@@ -1,1 +1,1 @@
-{}
+{"180470172660793344":{"points":0,"rank":"Neutral"},"192711002117242880":{"points":0,"rank":0}}
