+++ server/middleware/index.js
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+
+export default function() {
+    var router = Router();
+
+    
+
+    return router;
+}
